Separate stories into features, add playground.
